+++ src/database/migrations/0109-add-any-to-permission_grant_entity_type_t.sql
@@ -1,0 +1,1 @@
+ALTER TYPE permission_grant_entity_type_t ADD VALUE IF NOT EXISTS 'any';